# target/classes/yaml/swagger..yaml
paths:
  /nuova-convenzione:
    post:
      operationId: nuova-convenzione
      summary: 'creauna nuova convenzione'
      responses:
        '200':
          description: 'Ritorna successo in caso di avvenuta creazione della convenzione'
          schema:
            x-oad-type: 'null'
            type: 'null'
          x-oad-type: response
      parameters:
        -
          name: NDG
          in: body
          required: true
          schema:
            x-oad-type: string
            type: string
          x-oad-type: parameter
        -
          name: codiceFICS
          in: path
          required: false
          type: string
          x-oad-type: parameter
        -
          name: ABI
          in: body
          required: true
          schema:
            x-oad-type: string
            type: string
          x-oad-type: parameter
        -
          name: tipoAccesso
          in: body
          required: true
          schema:
            x-oad-type: string
            type: string
          x-oad-type: parameter
        -
          name: documenti
          in: body
          required: true
          schema:
            x-oad-type: array
            type: array
            items:
              x-oad-type: object
              type: object
          x-oad-type: parameter
    x-oad-type: operation
definitions:

  Convenzione:
    x-oad-type: object
    type: object
    properties:
      NDG:
        x-oad-type: string
        type: string
      codiceFIcs:
        x-oad-type: string
        type: string
      ABI:
        x-oad-type: string
        type: string
      tipoAccesso:
        x-oad-type: string
        type: string
      documentiCaricati:
        x-oad-type: array
        type: array
        description: 'documenti necessari per la creazione del convenzionamento'
        items:
          x-oad-type: string
          type: string
          description: 'codice documento caricato su filenet'
info:
  title: ConvenzioneAPI
  version: '1.0'
  description: 'API che crea/ricerca/edita un convenzionamento'
host: esempio.com
basePath: /convenzione
swagger: '2.0'