# target/generated-sources/src/main/java/it/prova/swagger/api/NuovaConvenzioneApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package it.prova.swagger.api;

import java.util.List;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-06T12:59:45.556+02:00")

@Api(value = "nuova-convenzione", description = "the nuova-convenzione API")
@RequestMapping(value = "/convenzione")
public interface NuovaConvenzioneApi {

    @ApiOperation(value = "creauna nuova convenzione", nickname = "nuovaConvenzione", notes = "", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Ritorna successo in caso di avvenuta creazione della convenzione") })
    @RequestMapping(value = "/nuova-convenzione",
        method = RequestMethod.POST)
    default ResponseEntity<Void> nuovaConvenzione(@ApiParam(value = "" ,required=true )  @Valid @RequestBody String NDG,@ApiParam(value = "",required=true) @PathVariable("codiceFICS") String codiceFICS,@ApiParam(value = "" ,required=true )  @Valid @RequestBody String ABI,@ApiParam(value = "" ,required=true )  @Valid @RequestBody String tipoAccesso,@ApiParam(value = "" ,required=true )  @Valid @RequestBody List<Object> documenti) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default NuovaConvenzioneApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}